implemented /block/transaction handler (it is suboptimal)
