+++ chain/rosettarpc/src/models.rs
@@ -101,7 +101,11 @@ pub(crate) struct Amount {
 }
 
 /// Blocks contain an array of Transactions that occurred at a particular
-/// BlockIdentifier. A hard requirement for blocks returned by Rosetta implementations is that they MUST be _inalterable_: once a client has requested and received a block identified by a specific BlockIndentifier, all future calls for that same BlockIdentifier must return the same block contents.
+/// BlockIdentifier. A hard requirement for blocks returned by Rosetta
+/// implementations is that they MUST be _inalterable_: once a client has
+/// requested and received a block identified by a specific BlockIndentifier,
+/// all future calls for that same BlockIdentifier must return the same block
+/// contents.
 #[derive(Debug, Clone, PartialEq, serde::Serialize, serde::Deserialize, Apiv2Schema)]
 pub(crate) struct Block {
     pub block_identifier: BlockIdentifier,
@@ -512,7 +516,11 @@ pub(crate) enum SyncStage {
 #[derive(Debug, Clone, PartialEq, serde::Serialize, serde::Deserialize, Apiv2Schema)]
 pub(crate) struct SyncStatus {
     pub current_index: i64,
+
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub target_index: Option<i64>,
+
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub stage: Option<SyncStage>,
 }
